Make tabObjects separated submodules
